requirements.txt
streamlit==1.31.1
streamlit-bokeh-events==0.1.2
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
langchain==0.0.352
langchain-community==0.0.13
langchain-core==0.1.23
langchain-openai==0.0.6
typing_extensions==4.9.0
openai==1.12.0
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0

chroma-hnswlib==0.7.3
chromadb==0.3.29
